# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CSV Analyser API",
        "description": "High-performance FastAPI backend for CSV data analysis, Open Icecat product enrichment, preview, and visualizations with MongoDB persistence.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/csv/enrich": {
            "post": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Enrich Csv Dataset",
                "description": "Upload a CSV file, enrich all products with Open Icecat technical specs & media,\nsave the newly generated enriched CSV, and return enrichment statistics and preview.",
                "operationId": "enrich_csv_dataset_api_csv_enrich_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_enrich_csv_dataset_api_csv_enrich_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EnrichmentSummary"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/verify": {
            "post": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Verify Csv Dataset",
                "description": "Upload a CSV, audit each product row against live Google/DuckDuckGo web index data,\ncalculate data closeness/fidelity scores, detect discrepancies, and return comprehensive insights.",
                "operationId": "verify_csv_dataset_api_csv_verify_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_verify_csv_dataset_api_csv_verify_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/{file_id}/download-verified": {
            "get": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Download Verified Csv",
                "description": "Download the verified CSV file containing web closeness scores, verification status, and audit insights.",
                "operationId": "download_verified_csv_api_csv__file_id__download_verified_get",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/{file_id}/download-enriched": {
            "get": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Download Enriched Csv",
                "description": "Download the generated enriched CSV file with all Open Icecat and technical parameters.",
                "operationId": "download_enriched_csv_api_csv__file_id__download_enriched_get",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/enrich/stream": {
            "post": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Stream Enriched Csv",
                "description": "Directly streams back the enriched CSV file as an attachment without requiring a secondary download.",
                "operationId": "stream_enriched_csv_api_csv_enrich_stream_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_stream_enriched_csv_api_csv_enrich_stream_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/verify/single": {
            "post": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Verify Single Product",
                "description": "Audit and cross-verify a single Brand + Model against live Google/DuckDuckGo web index data.",
                "operationId": "verify_single_product_api_csv_verify_single_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SingleVerificationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleVerificationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/enrich/item": {
            "post": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Enrich Single Product",
                "description": "Query Open Icecat directly for a single Brand + Model Code combination.",
                "operationId": "enrich_single_product_api_csv_enrich_item_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SingleEnrichmentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleEnrichmentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Body_enrich_csv_dataset_api_csv_enrich_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File",
                        "description": "CSV file with product codes/brands"
                    },
                    "brand_column": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Brand Column",
                        "description": "Optional column name for Brand"
                    },
                    "model_column": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model Column",
                        "description": "Optional column name for Model/Product Code"
                    },
                    "description_column": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description Column",
                        "description": "Optional column name for Product Description/Title"
                    },
                    "ean_column": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ean Column",
                        "description": "Optional column name for EAN/GTIN"
                    },
                    "include_web_scraping": {
                        "type": "boolean",
                        "title": "Include Web Scraping",
                        "description": "Enable web scraping fallback for missing data",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_enrich_csv_dataset_api_csv_enrich_post"
            },
            "Body_stream_enriched_csv_api_csv_enrich_stream_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File"
                    },
                    "brand_column": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Brand Column"
                    },
                    "model_column": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model Column"
                    },
                    "include_web_scraping": {
                        "type": "boolean",
                        "title": "Include Web Scraping",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_stream_enriched_csv_api_csv_enrich_stream_post"
            },
...
            "Body_verify_csv_dataset_api_csv_verify_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File",
                        "description": "CSV file to cross-verify against live DuckDuckGo web data"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_verify_csv_dataset_api_csv_verify_post"
            },
...
            "EnrichmentSummary": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id"
                    },
                    "original_filename": {
                        "type": "string",
                        "title": "Original Filename"
                    },
                    "enriched_filename": {
                        "type": "string",
                        "title": "Enriched Filename"
                    },
                    "total_rows": {
                        "type": "integer",
                        "title": "Total Rows"
                    },
                    "matched_icecat_count": {
                        "type": "integer",
                        "title": "Matched Icecat Count"
                    },
                    "web_enriched_count": {
                        "type": "integer",
                        "title": "Web Enriched Count",
                        "default": 0
                    },
                    "total_enriched_count": {
                        "type": "integer",
                        "title": "Total Enriched Count",
                        "default": 0
                    },
                    "match_rate_percentage": {
                        "type": "number",
                        "title": "Match Rate Percentage"
                    },
                    "brand_column_used": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Brand Column Used"
                    },
                    "model_column_used": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model Column Used"
                    },
                    "description_column_used": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description Column Used"
                    },
                    "ean_column_used": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ean Column Used"
                    },
                    "new_columns_added": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "New Columns Added"
                    },
                    "download_url": {
                        "type": "string",
                        "title": "Download Url"
                    },
                    "preview_rows": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Preview Rows",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "file_id",
                    "original_filename",
                    "enriched_filename",
                    "total_rows",
                    "matched_icecat_count",
                    "match_rate_percentage",
                    "new_columns_added",
                    "download_url"
                ],
                "title": "EnrichmentSummary"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "VerificationResponse": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id"
                    },
                    "original_filename": {
                        "type": "string",
                        "title": "Original Filename"
                    },
                    "verified_filename": {
                        "type": "string",
                        "title": "Verified Filename"
                    },
                    "dataset_insights": {
                        "$ref": "#/components/schemas/DatasetInsights"
                    },
                    "preview_rows": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Preview Rows",
                        "default": []
                    },
                    "download_url": {
                        "type": "string",
                        "title": "Download Url"
                    }
                },
                "type": "object",
                "required": [
                    "file_id",
                    "original_filename",
                    "verified_filename",
                    "dataset_insights",
                    "download_url"
                ],
                "title": "VerificationResponse"
            }
        }
    }
}